fixed CORS proxy to support non-json responses
the proxy tried to decode the forwarded response from json before sending it, but this assumed that all responses were in json format, and triggered an error if it got anything else, like raw text. now it just grabs the text and sends that without parsing it, hopefully working with any response type.